More

+++ tools/eoc/driver.py
@@ -288,7 +288,7 @@ class Pipeline:
             shutil.copyfile(self.plugin_generated("trim_defs/trim_gen.eo"), output_file)
             return output_file
         finally:
-            if temp_trim.exists():
+            if temp_trim.exists() and os.environ.get("EOC_KEEP_TEMP_TRIM", "0") == "0":
                 temp_trim.unlink()
 
     def desugar(
